==> 05_protein_keyword_featurization.ipynb <==
# Purpose

The purpose of this notebook is to turn the keywords gathered in the protein table into features for the train set.

Lets write the "learning ready" features I'm actually going to use in the model into another sqlite database.

That way I can keep my "raw" data separate from my "prepared" data.

import sqlite3
import pandas as pd
import numpy as np
import nltk
from tqdm import tqdm_notebook

conn = sqlite3.connect("human_protein.db")
cur = conn.cursor()
tables = cur.execute("SELECT name FROM sqlite_master WHERE type='table' ORDER BY name;").fetchall()
print(tables)
protein_df = pd.read_sql('SELECT * FROM protein',con=conn)
conn.close()
display(protein_df.head())

# Parse keywords to columns
protein_df.loc[:,'keyword_split'] = protein_df['keywords'].str.split(',').apply(frozenset)
for keyword in tqdm_notebook(frozenset.union(*protein_df.keyword_split)):
    protein_df[keyword.strip()] = protein_df.apply(lambda x: int(keyword in x.keyword_split),axis=1)

protein_df_keywords = protein_df.drop(["keywords","keyword_split"],axis=1)
protein_df_keywords.drop(['accession','sequence'],axis=1,inplace=True)
protein_df_keywords.head()

protein_df_keywords_no_empty = protein_df_keywords[[x for x in protein_df_keywords.columns if len(x)>0]]

conn_train = sqlite3.connect("protein_training.db")
protein_df_keywords_no_empty.to_sql('protein_keywords', con=conn_train, index=False, if_exists='replace')
conn_train.close()

#disease_df = pd.read_sql('SELECT * FROM protein_diseases',con=conn)
#display(disease_df.head())
#disease_df.groupby('disease').count().head()

Judging by how sparse disease is - I don't think it will contain enough information without some sort of parsing to be useful. Skip for now.
