==> 09_train_set_gen.ipynb <==
import sqlite3
import pandas as pd
import numpy as np
from tqdm import tqdm_notebook
from matplotlib import pyplot as plt
import seaborn as sns
import nltk
import sklearn
from Bio.SeqUtils.ProtParam import ProteinAnalysis
import math

conn = sqlite3.connect("protein_training.db")
proteins_df = pd.read_sql('SELECT * from protein_ngram LIMIT 1000',con=conn)
keyword_df = pd.read_sql('SELECT * from protein_keywords LIMIT 1000',con=conn)
seqfeat_df = pd.read_sql('SELECT * from protein_seq_features LIMIT 1000',con=conn)
subloc_df = pd.read_sql('SELECT * from protein_subloc LIMIT 1000',con=conn)
tissue_df = pd.read_sql('SELECT * from protein_tissue_exp_keywords LIMIT 1000',con=conn)
conn.close()

for df in [proteins_df, seqfeat_df, subloc_df, tissue_df]:
    display(df.head())

# Lets build the train set and the sample set

Train is all of the n-grams for the proteins with some amyloid

Sample are proteins with no amyloid at all

# List of proteins
conn = sqlite3.connect("protein_training.db")
proteins = pd.read_sql("SELECT DISTINCT protein FROM protein_ngram", con=conn)
conn.close()

conn = sqlite3.connect("protein_training.db")
subloc_prots = pd.read_sql("SELECT DISTINCT protein FROM protein_subloc", con=conn)
tissue_prots = pd.read_sql("SELECT DISTINCT protein FROM protein_tissue_exp_keywords", con=conn)
conn.close()

# Proteins not in subloc keywords
print(proteins[proteins.protein.isin(subloc_prots.protein)==False].shape[0])
# Proteins not in the tissue keywords
print(proteins[proteins.protein.isin(tissue_prots.protein)==False].shape[0])

# Manually delete table before running
conn = sqlite3.connect("protein_training.db")
dfs = ["protein_ngram", "protein_seq_features"]
all_cols = []
col_count_dict = {}
for df in dfs:
    cols = [x[1] for x in conn.execute("PRAGMA table_info("+df+");") if x[1]!="accession"]
    for col in cols:
        if col not in col_count_dict:
            col_count_dict[col] = 1
        else:
            col_count_dict[col] += 1
        
    all_cols.extend(cols)

conn.close()

col_df = pd.DataFrame.from_dict(col_count_dict,orient="index")
col_df[col_df.iloc[:,0]>1]
len(col_df)

# Put all the frames together
def chunks(l, n):
    """Yield successive n-sized chunks from l."""
    for i in range(0, len(l), n):
        yield tuple(l[i:i + n])

def merger(cols, proteins, conn):
    """Given protein name, construct the full feature arrays for that protein
    and write to the sqlite repo."""
    dfs = ["protein_ngram", "protein_seq_features"]
    final_frame = pd.DataFrame(list(proteins), columns=['protein'])
    for df in dfs:
        sub_set = pd.read_sql("SELECT * FROM "+df+" WHERE `protein` IN "+str(proteins), con=conn)
        if sub_set.shape[0] > 0:
            final_frame = final_frame.merge(sub_set, on='protein', how='outer')
    # Backfill missing cols
    for col in cols:
        if col not in final_frame.columns:
            final_frame.loc[:,col] = 0
    # Drop certain columns
    drop_cols = ["accession"]
    final_frame.drop(drop_cols,axis=1, inplace=True)
    return final_frame.fillna(0)

def protein_writer(cols, proteins, chunk_size, conn):
    prot_chunks = chunks(proteins, chunk_size)
    for chunk in tqdm_notebook(prot_chunks):
        frame = merger(cols, chunk, conn)
        frame.to_sql("train",con=conn, if_exists="append", index=False)

conn = sqlite3.connect("protein_training.db")
protein_writer(all_cols, proteins.protein.tolist(), 5000, conn)
conn.close()

s_its = [1.99, 2.66, 5.61, 8.46, 26.66, 49.14]
c_size = [10, 50, 500, 1000, 5000, 10000]
n = len(proteins.protein.tolist())

sits_df = pd.DataFrame(list(zip(s_its, c_size)), columns=["sec_per_it","chunk_size"])

sits_df.loc[:,'total_its'] = n / sits_df.chunk_size
sits_df.loc[:,'time_to_completion'] = sits_df['sec_per_it'] * sits_df['total_its']
sits_df

# Add table of amyloid or not (y-variable)
# def amyloidValues(protein_list, chunk_size):
#     prot_chunks = chunks(proteins, chunk_size)
#     for chunk in tqdm_notebook(prot_chunks):
conn = sqlite3.connect("protein_training.db")
amyloid_prots = pd.read_sql(
    "SELECT DISTINCT protein \
    FROM protein_ngram \
    WHERE \
        amyloid_1 = 1 OR \
        amyloid_2 = 1 OR \
        amyloid_3 = 1 OR \
        amyloid_4 = 1 OR \
        amyloid_5 = 1",
    con=conn
)
conn.close()
        #frame = merger(cols, chunk, conn)
        #frame.to_sql("train",con=conn, if_exists="append", index=False)

amyloid_prots.loc[:,'amyloid'] = 1
print(amyloid_prots.shape)
amyloid_prots.head()

non_amy_prots = proteins[proteins.protein.isin(amyloid_prots.protein.values.tolist())==False]
#proteins
non_amy_prots.loc[:,'amyloid'] = 0

train_y = pd.concat([amyloid_prots, non_amy_prots])
print(train_y.shape[0]==proteins.shape[0])
print(train_y.shape[0])
train_y.sample(5)

conn = sqlite3.connect("protein_training.db")
train_y.to_sql('amyloid', con=conn, index=False, if_exists="replace")
conn.close()
