==> 11.20_FinalProject_Morawej.Rmd <==
---
title: "Social Determinants of Cesearean Section: Analysis of Results from the National Survey of Family Growth"
author: "Sabrina Morawej"
output: 
 html_document:
    toc: true
    toc_float: 
        collapsed: false
        smooth_scroll: true
    depth: 3 
    theme: united
    highlight: tango

---
```{r set-options, echo=FALSE, cache=FALSE, message = FALSE}
options(width = 400)

# packages
library(tidyverse)
library(ggplot2)
library(gtsummary)
library(cowplot)
library(ggpubr)
library(dplyr)
```  

***
### Overview
  Medically unnecessary Cesarean sections (C-sections) are associated with poor health outcomes for mother and baby. This project will explore several social predictors of C-sections.
  
  For this project, I conferred with Dr. Shimrit Keddem, PI at the Center for Health Equity, Research, and Promotion; she has used the NSFG in previous publications, and familiarized me with the National Survey of Family Growth which is a freely accessible dataset from the CDC to analyze my outcome. Dr. Keddem also helped me identify the revelant social predictors to explore this topic. Dr. Allison Curry, PI at the Center for Injury Research and Prevention helped me gather the background data on this topic and identify critical gaps in the literature. Finally, Professor Jessica Rast, STATA instructor and PhD Epidemiology candidate at Drexel University, taught me how to import the data into STATA so that it could then be imported into R.

[link to my repo](https://github.com/sem-22/BMIN503_Final_Project)

### Introduction 
   In 2020, about 26% of births in the United States ended in Cesarean delivery.[1] As C-section rates continue to rise in the United States, the Department of Health and Human Services announced one of the Healthy People 2030’s goals to reduce C-section rates among low-risk first births.[2] Compared to vaginal births, C-sections pose greater risks of complications at birth,[3] the postpartum period,[4] and negative long-term health outcomes.[5] These complications translate into high health care costs for patients, hospitals, and health insurance providers and diminished levels of maternal and infant health.[6–9]
  
  
   The issue of rising rates of medically unneccessary C-sections requires an interdisciplinary approach. The goal of public healh is to prevent negative healh outcomes. Healh care providers, government officials, insurance companies, and hospital systems all share this goal. Additionally, they share another goal which is to reduce unneccessary medical costs. An economic and policy analysis contribute to the understanding of this insidious trend. 

### Methods
  I will use the National Survey of Family Growth's (NSFG) 2017-2019 Female Pregnancy file for this analysis. The Female Pregnancy file is a subset of the NSFG that contains all relevant variables related to pregnancy and birth. The NSFG is a cross-sectional, nationally representative study conducted by the Centers for Disease Control and Prevention. For the purposes of this analysis, the analytic sample only includes first-births. The independent variables are prenatal care, age, insurance status, race/ethnicity, and smoking status during pregnancy. The dependent variable is C-section. Logistic regression will be used to explore the relationship between these social predictors and C-section.
  
   As of today, the NSFG does not have downloadable files available for use in RStudio; nevertheless, the data were imported into R by creating a .dta outfile in STATA and then read into R through the appropriate "Import Dataset" feature. Using the *write.csv* command, the data were saved to R.
   
   To clean this dataset, I am filtering the dataset by first pregnancy. The literature shows that the strongest predictor of C-section is having a previous C-section. By excluding all pregnancies but the first, I am eliminating a known confounder. I am also excluding any birth that resulted in twins, triplets, quadruplets, etc. because those pregnancies and births have additional risks outside the scope of this project. I am also defining my outcome variable as "c-section vs. vaginal birth" using the variable: *PREGEND1* (transformed into *birth.outcome*). The question asks each respondent: "How Pregnancy Ended - 1st mention". 
   
   I will perform logistic regression to compare odds of having a c-section vs. a vaginal birth; therefore, I excluded all other pregnancy outcomes. The social predictors I am including in this analysis are: maternal age, number of cigarettes smoked per day during pregnancy, race, Hispanic origin, prenatal care, and insurance status.
   
  *There is a typo in the upload of 2017-2019 data which says this is the 2015-2017 NSFG, but each observation has a "Calendar year when interview occurred" (intvwyear) variable and all observations occurred within the stated 2017-2019 study period. I also contacted the NCHS, and I was reaffirmed that it was a mistype on their end.

``` {r, eval = FALSE}
write.csv(nsfg, file = "nsfg.csv", row.names = FALSE, quote = FALSE)
```

``` {r}
# save datafile in a way that R will save it when it's time to knit
nsfg <- read.csv("nsfg.csv")
```


``` {r}

# WGT2017_2019: Final weight for the 2017-2019 NSFG; Applicable for all Respondents
# secu: Randomized version of the sampling error computational unit; : Applicable for all pregnancies (respondent-level variable)
# sest: Randomized version of the stratum; Applicable for all pregnancies (respondent-level variable)

# Clean the data
clean_nsfg <- nsfg %>%
    filter(pregordr == 1) %>% # filtering for responses about R's first pregnancy (all pregnancies after the first are excluded).
    select(PREGEND1, agepreg, npostsmk, race, hispanic, pncarewk, paydeliv, WGT2017_2019, secu, sest) %>% # selecting all relevant social predictors including my dependent variable.
    rename(birth.outcome = PREGEND1, mom.age = agepreg, cigs.per.day = npostsmk, prenatal.care = pncarewk, delivery.payment.type = paydeliv) %>% # giving vars common sense names
    mutate(birth.outcome = factor(birth.outcome, levels = c(5, 6), labels = c("C-section", "vaginal birth"))) %>% # by setting birth.outcome of first pregnancy to live c-section or live vaginal birth, I am creating a nominal, dichotomous outcome variable (which is what I need for logistic regression) 
    mutate(cigs.per.day = factor(cigs.per.day, levels = c(1, 2, 3, 4, 6), labels = c("about 1", "2-4", "5-14", "15-24", "35-44"))) %>% # removed the "25-34" category (5th level) because this response was empty
    mutate(hispanic = factor(hispanic, levels = c(1, 2), labels = c("Hispanic", "Non-hispanic"))) %>%
    mutate(race = factor(race, levels = c(1, 2, 3), labels = c("Black", "White", "Other"))) %>%
    mutate(prenatal.care = factor(prenatal.care, levels = c(1, 2, 3), labels = c("<13 weeks", ">14 weeks", "no prenatal care"))) %>%
    mutate(delivery.payment.type = factor(delivery.payment.type, levels = c(1, 2, 3, 4, 5), labels = c("own income only", "insurance only", "own income & insurance only", "medicaid/govt assistance", "all other combinations of payment")))

```

### Results

Table 1: Descriptive Statistics of Study Population
``` {r, message = FALSE, warning = FALSE}
library(tableone)

## Vector of variables to summarize
myVars <- c("birth.outcome", "prenatal.care", "delivery.payment.type", "mom.age", "cigs.per.day", "race", "hispanic")

## Vector of categorical variables that need transformation
catVars <- c("birth.outcome", "prenatal.care", "delivery.payment.type", "cigs.per.day", "race", "hispanic")

## Create a TableOne object
table1 <- CreateTableOne(vars = myVars, data = clean_nsfg, factorVars = catVars)
table1

```

Plots for bivariate relationships between each social predictor and birth outcome.
``` {r, message = FALSE, warning = FALSE}
# bar chart of women whose first birth was a C-section or vaginal birth by race (Black, white, and other)
clean_nsfg %>%
    drop_na(birth.outcome, race) %>%
    ggplot(aes(x = birth.outcome, fill = race)) +
    geom_bar(position = "dodge") +
    ggtitle("Birth Outcome by Race") +
    xlab("Birth Outcome") +
    ylab("Count")

# number of women whose first birth was a C-section or vaginal birth by Hispanic origin (Hispanic vs. non-Hispanic)
clean_nsfg %>%
    drop_na(birth.outcome, hispanic) %>%
    ggplot(aes(x = birth.outcome, fill = hispanic)) +
    geom_bar(position = "dodge") +
    ggtitle("Birth Outcome by Hispanic Origin") +
    xlab("Birth Outcome") +
    ylab("Count")

# number of women whose first birth was a C-section or vaginal birth by the number of cigarettes she smoked during pregnancy after she knew she was pregnant
clean_nsfg %>%
    drop_na(birth.outcome, cigs.per.day) %>%
    ggplot(aes(x = birth.outcome, fill = cigs.per.day)) +
    geom_bar(position = "dodge") +
    ggtitle("Birth Outcome by Amount Smoked during Pregnancy") +
    xlab("Birth Outcome") +
    ylab("Count")

# number of women whose first birth was a C-section or vaginal birth by the number of weeks pregnant at first prenatal care visit
clean_nsfg %>%
    drop_na(birth.outcome, prenatal.care) %>%
    ggplot(aes(x = birth.outcome, fill = prenatal.care)) +
    geom_bar(position = "dodge") +
    ggtitle("Birth Outcome by Number of Weeks Pregnant at First Prenatal Care Visit") +
    xlab("Birth Outcome") +
    ylab("Count")

# Maternal Age by C-section and vaginal birth
clean_nsfg %>%
     drop_na(birth.outcome, mom.age) %>%
    ggplot(aes(x = birth.outcome, y = mom.age)) +
    geom_boxplot(color = "black", fill = "blue") +
    ggtitle("Birth Outcome by Maternal Age") +
    xlab("Birth Outcome") +
    ylab("Maternal Age")

# number of women whose first birth was a C-section or vaginal birth by the type of payment method for the delivery
clean_nsfg %>%
    drop_na(birth.outcome, delivery.payment.type) %>%
    ggplot(aes(x = birth.outcome, fill = delivery.payment.type)) +
    geom_bar(position = "dodge") +
    ggtitle("Birth Outcome by Delivery Payment Type") +
    xlab("Birth Outcome") +
    ylab("Count")

```

The crude odds ratios for each social predictor: maternal age, number of cigarettes smoked per day during pregnancy, race, Hispanic origin, prenatal care, and insurance status
``` {r, message = FALSE, warning = FALSE}
# maternal age (I/R variable)
summary(glm(birth.outcome ~ mom.age, data = clean_nsfg, 
                  family = binomial()))
# cigarettes smoked per day after confirming pregnancy (ordinal variable)
summary(glm(birth.outcome ~ cigs.per.day, data = clean_nsfg, 
                  family = binomial())) 
# race (nominal variable)
summary(glm(birth.outcome ~ race, data = clean_nsfg, 
                  family = binomial()))
# Hispanic origin (nominal, dichotomous variable)
summary(glm(birth.outcome ~ hispanic, data = clean_nsfg, 
                  family = binomial()))
# prenatal care (nominal variable)
summary(glm(birth.outcome ~ prenatal.care, data = clean_nsfg, 
                  family = binomial()))
# payment type for delivery  (nominal variable)
summary(glm(birth.outcome ~ delivery.payment.type, data = clean_nsfg, 
                  family = binomial()))
```


Table 2. Unadjusted ORs from the univariate regressions of each predictor
```{r, echo=FALSE, warning = FALSE}
clean_nsfg %>%
  select(birth.outcome, mom.age, cigs.per.day, race, hispanic, prenatal.care, delivery.payment.type) %>%
  tbl_uvregression(
    method = glm,
    y = birth.outcome,
    method.args = list(family = binomial),
    exponentiate = TRUE,
    pvalue_fun = ~style_pvalue(.x, digits = 2)
  ) %>%
  add_global_p() %>%  # add global p-value 
  add_nevent() %>%    # add number of events of the outcome
  add_q() %>%         # adjusts global p-values for multiple testing
  bold_p() %>%        # bold p-values under a given threshold (default 0.05)
  bold_p(t = 0.05, q = TRUE) %>% # now bold q-values under the threshold of 0.10
  bold_labels()

```

The multivariate logistic regression with all predictors
``` {r, message = FALSE, warning = FALSE}
mlr <- birth.outcome.glm <- glm(birth.outcome ~ mom.age + cigs.per.day + race + hispanic + prenatal.care + delivery.payment.type, data = clean_nsfg, family = binomial())

```

Table 3. Adjusted ORs from the multiple logistic regressions of each predictor
``` {r, message = FALSE, warning = FALSE}
mlr %>%
  tbl_regression(exponentiate = TRUE) %>%
  bold_labels() %>%
  bold_p(t = 0.05)

```

### Conclusion
