==> 19_feature_importance_viz.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

This is an [R Markdown](http://rmarkdown.rstudio.com) Notebook. When you execute code within the notebook, the results appear beneath the code. 

Try executing this chunk by clicking the *Run* button within the chunk or by placing your cursor inside it and pressing *Ctrl+Shift+Enter*. 

```{r}
library(caret)
library(ggplot2)
library(dplyr)
library(tidyr)
library(reshape)
library(RColorBrewer)

cpal <- brewer.pal(n = 5, name = "BuGn")

load("D:/GitHub/BMIN503_Final_Project/c5tree_4.RData")
feat_imps <- varImp(model)$importance 
feat_imps$feature <- rownames(feat_imps)

colnames(feat_imps) <- c("Importance","Feature")

top_feat_imps <- feat_imps %>%
  filter(Importance >= 70)

feat_imp_plot <- ggplot(filter(feat_imps, Importance >0), aes(x=Feature[order(Feature, Importance, decreasing = TRUE)], y=Importance)) +
  geom_bar(stat="identity", fill="#018571") +
  scale_x_discrete(name="Feature") +
  scale_y_continuous(name="Importance", expand=c(0,0)) +
  coord_flip() +
  ggtitle(label="Top Features for C5 Tree") +
  theme_bw() +
  theme(axis.text.y = element_text(size=5)) + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))

top_feat_imp_plot <- ggplot(top_feat_imps, aes(x=Feature[order(Feature, Importance, decreasing = TRUE)], y=Importance)) +
  geom_bar(stat="identity", fill="#018571") +
  scale_x_discrete(name="Feature") +
  scale_y_continuous(name="Importance", expand=c(0,0)) +
  #coord_cartesian(ylim=c(70, 100)) +
  coord_flip(ylim=c(70,100)) +
  ggtitle(label="Top Features for C5 Tree") +
  theme_bw() +
  theme(axis.text.y = element_text(size=5)) + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))



ggsave(feat_imp_plot, filename = "all_tree_feats.png", height = 6, width=6, units="in", dpi=200)

ggsave(top_feat_imp_plot, filename = "top_tree_feats.png", height = 6, width=6, units="in", dpi=200)
```
